+++ src/renderer/components/views/SettingsViews/VersionSettingsView.tsx
@@ -62,15 +62,16 @@ export default function VersionSettingsView() {
                       install: { active: true, callback: reload }
                     }} className={styles.card} />)}
                 </div>;
-              }} className={[styles.scrollable, "tabs"].join(' ')} />
+              }} className={[styles.scrollable, 'tabs'].join(' ')} />
 
 
             });
 
           });
           Promise.all(construct).then(() => {
-            resolve(<Tabs items={tabList} className={[styles.VersionTool,styles.scrollable, 'scrollable HideOperation'].join(' ')}
-                          type={'card'} centered  />);
+            resolve(<Tabs items={tabList}
+                          className={[styles.VersionTool, styles.scrollable, 'scrollable HideOperation'].join(' ')}
+                          type={'card'} centered />);
           });
         }).catch(err => console.error(err));
     });
@@ -80,7 +81,7 @@ export default function VersionSettingsView() {
         <PathInput callback={() => reload()} />
       </div>
     );
-  }} className={[defaultStyle.View, styles.View, "versionSettings"].join(' ')} style={undefined} />;
+  }} className={[defaultStyle.View, styles.View, 'versionSettings'].join(' ')} style={undefined} />;
 }
 
 function PathInput({ callback }: { callback: () => any }): JSX.Element {
@@ -124,7 +125,9 @@ function PathInput({ callback }: { callback: () => any }): JSX.Element {
       const saved: string = await window.electron.ipcRenderer.invoke('GameEngine:getRootPath', {});
       const defaultPath: string = await window.electron.ipcRenderer.invoke('GameEngine:getDefaultRootPath', {});
       return (
-        <LabeledInput label={'minecraft folder path'} input={
+        <LabeledInput label={'minecraft folder path'}
+
+        >
           <div className={styles.pathSelector}>
             {state !== states.normal && <Icon icon={getIcon()} className={styles.icon} />}
             <AutoComplete status={state === states.error ? 'error' : undefined}
@@ -140,7 +143,7 @@ function PathInput({ callback }: { callback: () => any }): JSX.Element {
             </AutoComplete>
 
           </div>
-        } />);
+        </LabeledInput>);
     }
     } className={styles.pathSelector} style={undefined} />
   );